+++ examples/bzlmod/entry_points/BUILD.bazel
@@ -1,15 +1,6 @@
 load("@python_versions//3.9:defs.bzl", py_console_script_binary_3_9 = "py_console_script_binary")
 load("@rules_python//python:defs.bzl", "py_test")
-load("@rules_python//python/entry_point:py_console_script_binary.bzl", "py_console_script_binary")
-
-# We can specifically request the py_console_script_binary to be running with Python 3.9.
-py_console_script_binary_3_9(
-    name = "yamllint",
-    pkg = "@pip//yamllint:pkg",
-    # yamllint does not have any other scripts except 'yamllint' so the
-    # user does not have to specify which console script we should chose from
-    # the package.
-)
+load("@rules_python//python/entry_points:py_console_script_binary.bzl", "py_console_script_binary")
 
 # Here we are running the default version.
 py_console_script_binary(
@@ -22,6 +13,15 @@ py_console_script_binary(
         # One can add extra dependencies to the entry point.
         "@pip//pylint_print",
     ],
+)
+
+# We can specifically request the py_console_script_binary to be running with Python 3.9.
+py_console_script_binary_3_9(
+    name = "yamllint",
+    pkg = "@pip//yamllint:pkg",
+    # yamllint does not have any other scripts except 'yamllint' so the
+    # user does not have to specify which console script we should chose from
+    # the package.
 )
 
 py_test(
