+++ examples/bzlmod/entry_points/test_entry_point.py
