+++ python/BUILD.bazel
@@ -36,7 +36,7 @@ filegroup(
         "//python/cc:distribution",
         "//python/config_settings:distribution",
         "//python/constraints:distribution",
-        "//python/entry_point:distribution",
+        "//python/entry_points:distribution",
         "//python/private:distribution",
         "//python/runfiles:distribution",
     ],
