+++ python/entry_points/BUILD.bazel
@@ -13,8 +13,6 @@
 # limitations under the License.
 
 load("@bazel_skylib//:bzl_library.bzl", "bzl_library")
-load("//python:py_binary.bzl", "py_binary")
-load("//python:py_library.bzl", "py_library")
 
 exports_files(
     [
@@ -23,41 +21,19 @@ exports_files(
     visibility = ["//docs:__subpackages__"],
 )
 
-py_binary(
-    name = "generator_bin",
-    srcs = ["generator.py"],
-    main = "generator.py",
-    visibility = [
-        "//visibility:public",
-    ],
-)
-
-py_library(
-    name = "generator",
-    srcs = ["generator.py"],
-    imports = ["../.."],
-    visibility = [
-        "//tests/entry_points:__pkg__",
-    ],
-)
-
 bzl_library(
     name = "py_console_script_binary_bzl",
     srcs = [":py_console_script_binary.bzl"],
     visibility = ["//visibility:public"],
     deps = [
-        "//python:py_binary_bzl",
-        "//python/entry_point/private:py_entry_point_gen.bzl",
+        "//python/private:py_console_script_binary_bzl",
     ],
 )
 
 filegroup(
     name = "distribution",
     srcs = glob([
-        "*.py",
         "*.bzl",
-    ]) + [
-        "//python/entry_point/private:distribution",
-    ],
+    ]),
     visibility = ["//python:__subpackages__"],
 )
