+++ python/pip_install/pip_repository_requirements_bzlmod.bzl.tmpl
@@ -33,7 +33,7 @@ def entry_point(pkg, script = None):
     fail("""Please replace this instance of entry_point with the following:
 
 ```
-load("@rules_python//python/entry_point:py_console_script_binary.bzl", "entry_point")
+load("@rules_python//python/entry_points:py_console_script_binary.bzl", "entry_point")
 
 entry_point(
     name = "{pkg}",
