+++ python/private/py_console_script_gen.py
