+++ tests/BUILD.bazel
@@ -30,6 +30,6 @@ build_test(
         "//python:py_test_bzl",
         "//python/cc:py_cc_toolchain_bzl",
         "//python/cc:py_cc_toolchain_info_bzl",
-        "//python/entry_point:py_console_script_binary_bzl",
+        "//python/entry_points:py_console_script_binary_bzl",
     ],
 )
