+++ tests/entry_points/BUILD.bazel
@@ -15,10 +15,12 @@
 load("//python:py_test.bzl", "py_test")
 
 py_test(
-    name = "generator_test",
-    srcs = ["generator_test.py"],
+    name = "py_console_script_gen_test",
+    srcs = ["py_console_script_gen_test.py"],
     imports = ["../.."],
-    main = "generator_test.py",
+    main = "py_console_script_gen_test.py",
     visibility = ["//visibility:private"],
-    deps = ["//python/entry_point:generator"],
+    deps = [
+        "//python/private:py_console_script_gen_lib",
+    ],
 )
